requirements.txt
requests
git+https://github.com/Rapptz/discord.py
PyNaCl==1.3.0
dnspython==1.16.0
discord.py==1.5.1
async-timeout==3.0.1
random2